In progress fix memory/lifetime issues.
I figured out why everything is broken. It is because i was not considering
the lifetimes of the buffers that I was passing to write/read calls. I need
to go back through and refactor it taking this into account, making sure no
vars go out of scope. Another Rust W.